Refactor services.  Still needs more refactoring.
