+++ src/main/java/es/tidetim/Application.java
@@ -3,9 +3,11 @@ package es.tidetim;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 @EnableAutoConfiguration
+@ComponentScan
 public class Application {
 
     public static void main(String[] args) {
