+++ src/main/java/es/tidetim/tideengine/models/TideType.java
@@ -1,4 +1,4 @@
-package tideengine;
+package es.tidetim.tideengine.models;
 
 
 public enum TideType {
