--- src/main/java/tideengine/StationsService.java
@@ -1,7 +1,0 @@
-package tideengine;
-
-public interface StationsService {
-
-	TideStation getTideStation(String stationName);
-
-}
