+++ src/main/resources/application.yml
@@ -2,7 +2,7 @@ application:
   name: tides
 
 elasticsearch:
-  cluster: elasticsearch
+  cluster: mygovscot
   host: localhost
   port: 9300
   indexName: tides
